requirements.txt
markdown-it-py==3.0.0
mdurl==0.1.2
Pygments==2.19.1
PyYAML==6.0.2
rich==13.9.4
